fix: remove the daemon-disconnect banner
A socket drop painted a full-width red DAEMON DISCONNECTED strip above the
pane row. Rove keeps working with the daemon down and the reconnect loop
recovers most drops in well under a second, so the alert interrupted with
nothing to act on — and reflowed the whole window on each blip. The skew
banner stays: a stale daemon build persists until someone restarts it.

Follows the dead code out: DaemonDownBanner, useDaemonDown, the
workspace.daemonDown.* and automations.daemonUnreachable keys in both
locales. connectionStateSignal() itself is untouched.
